requirements.txt
fastapi
pymongo
uvicorn
jinja2
numpy
pandas
python-jose[cryptography]
passlib[bcrypt]
starlette
motor
jwt
python-multipart
cryptography
requests